refactor(jetsocat): split the doctor module
The module was growing above the 1000 lines of code. I’m also currently
adding a set of diagnostics using schannel, and it will likely add
~350 lines of code. For easier navigation, it is split into multiple
submodules.
